Backend/FASTAPI-QuanLyKho/requirements.txt
fastapi==0.100.0
Pillow==9.5.0
pydantic==1.10.7
PyJWT==1.7.1
python-decouple==3.8
SQLAlchemy==2.0.13
uvicorn
mysql